requirements.txt
opencv-python==4.8.1.78
deepface==0.0.79
tensorflow==2.13.0
pandas==2.0.3
numpy==1.24.3
pyserial==3.5
flask==2.3.3
flask-socketio==5.3.5
python-socketio==5.8.0
pyttsx3==2.90
scikit-learn==1.3.0
pillow==10.0.0